+++ app/spike/internal/cmd/delete.go
@@ -6,7 +6,6 @@ package cmd
 
 import (
 	"fmt"
-
 	"strconv"
 	"strings"